+++ app/src/main/java/com/github/nakawai/newsreader/domain/model/NYTimesModel.kt
@@ -1,11 +1,11 @@
-package com.github.nakawai.newsreader.data
+package com.github.nakawai.newsreader.domain.model
 
 import androidx.lifecycle.LiveData
 import com.github.nakawai.newsreader.domain.entities.Section
 import com.github.nakawai.newsreader.domain.entities.Story
 import com.github.nakawai.newsreader.domain.entities.StoryUrl
 
-interface Repository {
+interface NYTimesModel {
     suspend fun loadNewsFeed(section: Section, forceReload: Boolean): List<Story>
     fun updateStoryReadState(storyUrl: StoryUrl, read: Boolean)
     fun observeArticlesBySection(section: Section): LiveData<List<Story>>
